pyproject.toml
[project]
name = "task11-mcp-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-core>=2.25.1",
    "google-generativeai>=0.8.5",
    "mcp>=1.10.1",
    "msal>=1.32.3",
    "requests>=2.32.4",
]
